clean up all sidebar ai summary references and refactor kea logic
