+++ frontend/src/scenes/session-recordings/player/sidebar/PlayerSidebarTab.tsx
@@ -8,7 +8,6 @@ import { SessionRecordingSidebarTab } from '~/types'
 import { PlayerSidebarLinkedIssuesTab } from './PlayerSidebarLinkedIssuesTab'
 import { playerSidebarLogic } from './playerSidebarLogic'
 import { PlayerSidebarOverviewTab } from './PlayerSidebarOverviewTab'
-import { PlayerSidebarSessionSummaryTab } from './PlayerSidebarSessionSummaryTab'
 
 export function PlayerSidebarTab(): JSX.Element | null {
     const { activeTab } = useValues(playerSidebarLogic)
@@ -20,8 +19,6 @@ export function PlayerSidebarTab(): JSX.Element | null {
             return <PlayerInspector />
         case SessionRecordingSidebarTab.NETWORK_WATERFALL:
             return <NetworkView />
-        case SessionRecordingSidebarTab.SESSION_SUMMARY:
-            return <PlayerSidebarSessionSummaryTab />
         case SessionRecordingSidebarTab.LINKED_ISSUES:
             return <PlayerSidebarLinkedIssuesTab />
         default:
